feat(gpu): glyph-atlas batched text
The profiled ~40 ms frames spend their largest single share in text:
per-string cached textures (one pipeline bind + draw each, ~1300 direct
draws), a text-cache hash walk per drawn string per frame, and texture
create/destroy churn for every changing string. This routes text
through the phase-1 instanced batcher:

Tier 1 - per-glyph quads (bitmap fonts, opaque draws, the default path):
- sdl_gpu_text.c rasterizes each (font, character) ONCE via
  sdl_text_glyph.c into the shared atlas and draws strings as one
  GPU_FX_MODE_PLAIN instance per glyph, colored per instance.
- No sdl_tx_load call at all: no cache entry, no hash walk, no
  per-string texture, NOCACHE overlays included.
- All-or-nothing emission (glyph readiness and instance capacity are
  verified before the first quad); glyph masks revalidate against the
  font's raw pointers, surviving the scale>1 PNG font swap.
- Opaque-only by design: all ink of a pass is one opaque color, so
  overlapping ink (derived _shaded/_framed fonts) writes identical
  pixels in any order - bit-exact per tests/test_text_compare.c.

Tier 2 - cached-string quads (TTF strings, alpha fades, fallbacks):
- sdl_maketext_gpu now places composed string pixels into the shared
  atlas when the batcher is up (standalone texture otherwise or when
  oversized); eviction returns the region (previous commit).
- sdl_drawtext_alpha draws cached entries as ONE batched plain quad
  (exact for every case - the pre-composed pixels blend once) and only
  falls back to the direct per-draw pipeline when the batch is
  unavailable or a non-standard blend mode is active; the direct text
  draw is now atlas-aware.

Both tiers ride the existing gpu_shader_effects flag: flag off = the
exact pre-change path. ASTONIA_GPU_STATS now reports text runs, glyph
quads and fallbacks per frame.

+++ src/sdl/sdl_text_glyph.c
@@ -88,8 +88,8 @@ int text_glyph_rasterize(
 	out->w = max_x + 1;
 	out->h = max_y + 1;
 	for (y = 0; y < out->h; y++) {
-		memmove(
-		    buffer + (size_t)y * out->w, buffer + (size_t)y * TEXT_GLYPH_MAX_DIM, (size_t)out->w * sizeof(uint32_t));
+		memmove(buffer + (size_t)y * (size_t)out->w, buffer + (size_t)y * TEXT_GLYPH_MAX_DIM,
+		    (size_t)out->w * sizeof(uint32_t));
 	}
 	out->pixel = buffer;
 	return 1;
